Added BathroomData, ReviewData, Coordinates, and ServerInterfacer classes.
BathroomData and ReviewData just have constructors, getters, and a
method that calls the ServerInterfacer's getReviews() and
markAsHelpful() methods respectively.

ServerInterfacer is entirely unimplemented stubs of the methods we
talked about in the last meeting.

Coordinates is just a coordinates tuple class. It has getters and a
constructor but no distance calculator yet.